+++ backend/test/python/data/d1.py
@@ -1,0 +1,11 @@
+
+
+from tools import *
+
+def test_circle():
+    tri_circle_multi()
+    match_multi()
+    table_multi()   
+    table_player_multi()
+
+test_circle()
